+++ ChiTech/ChiMesh/SweepUtilities/FLUDS/AUX_FLUDS.cc
@@ -72,7 +72,7 @@ OutgoingPsi(int cell_so_index, int outb_face_counter,
       local_psi_Gn_block_strideG[fc]*n +
         so_cell_outb_face_slot_indices[cell_so_index][outb_face_counter] *
         local_psi_stride[fc] *G +
-      face_dof*G;
+      (size_t) face_dof*G;
 
     return &(ref_local_psi->operator[](fc))[index];
   }
@@ -82,7 +82,7 @@ OutgoingPsi(int cell_so_index, int outb_face_counter,
       delayed_local_psi_Gn_block_strideG*n +
         so_cell_outb_face_slot_indices[cell_so_index][outb_face_counter] *
         delayed_local_psi_stride *G +
-      face_dof*G;
+      (size_t) face_dof*G;
 
     return &ref_delayed_local_psi->operator[](index);
   }
@@ -143,7 +143,7 @@ UpwindPsi(int cell_so_index, int inc_face_counter,
       local_psi_Gn_block_strideG[fc]*n +
         so_cell_inco_face_dof_indices[cell_so_index][inc_face_counter].slot_address *
         local_psi_stride[fc] *G +
-        so_cell_inco_face_dof_indices[cell_so_index][inc_face_counter].
+        (size_t) so_cell_inco_face_dof_indices[cell_so_index][inc_face_counter].
         upwind_dof_mapping[face_dof] *G + g;
 
     return &(ref_local_psi->operator[](fc))[index];
@@ -154,7 +154,7 @@ UpwindPsi(int cell_so_index, int inc_face_counter,
       delayed_local_psi_Gn_block_strideG*n +
         so_cell_inco_face_dof_indices[cell_so_index][inc_face_counter].slot_address *
         delayed_local_psi_stride *G +
-        so_cell_inco_face_dof_indices[cell_so_index][inc_face_counter].
+        (size_t) so_cell_inco_face_dof_indices[cell_so_index][inc_face_counter].
         upwind_dof_mapping[face_dof] *G + g;
 
     return &ref_delayed_local_psi_old->operator[](index);
